added delete message feature for file message

+++ src/controllers/messageController.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, Response } from "express";
 import { prisma } from "../lib/prisma";
 import { renameSync } from "fs";
 import { BASE_URL_SERVER } from "../lib/BASE_URL";
-import { getObjectURL, putObject } from "../aws";
+import { deleteObject, getObjectURL, putObject } from "../aws";
 const path = require("path");
 import axios from "axios";
 
@@ -85,6 +85,13 @@ export async function getMessagesController(
       },
     });
 
+    for (let message of messages) {
+      if (message.type === 'file') {
+        const imageUrl = await getObjectURL(message.message);
+        message.message = imageUrl;
+      }
+    }
+
     res.status(200).json(messages);
     return;
   } catch (e) {
@@ -92,6 +99,60 @@ export async function getMessagesController(
     console.log(`Error in postMessageController ${e}`);
     next(e);
     return;
+  }
+}
+
+export async function uploadFileController(
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) {
+  try {
+    const file = req.file;
+    const senderId = req.query.senderId?.toString();
+    const receiverId = req.query.receiverId?.toString();
+    const type = req.file?.mimetype || "";
+    const filePath = req.file?.path || "";
+
+    if (!file) return res.status(400).json({ error: `File is required` });
+    if (!senderId) return res.status(400).json({ error: `senderId is required` });
+    if (!receiverId) return res.status(400).json({ error: `receiverId is required` });
+
+    const filename = file?.filename || "";
+    const key = `uploads/file-message/${Date.now()}-${filename}`;
+
+    const url = await putObject({
+      key,
+      contentType: type,
+    });
+
+    const response = await axios.put(url, file.buffer, {
+      headers: {
+        "Content-Type": type,
+      },
+    });
+    // console.log(response);
+
+    const message = await prisma.message.create({
+      data: {
+        senderUser: {
+          connect: { id: senderId },
+        },
+        receiverUser: {
+          connect: { id: receiverId },
+        },
+        message: key,
+        type: "file",
+      },
+    });
+    const imageUrl = await getObjectURL(key);
+
+    return res.status(201).json(imageUrl);
+  } catch (e) {
+    console.log(`Error in uploadFileController ${e}`);
+    return res
+      .status(500)
+      .json({ error: `Error in uploadFileController ${e}` });
   }
 }
 
@@ -107,8 +168,19 @@ export async function deleteMessageController(
       res.status(400).json({ error: "MessageId is required" });
       return;
     }
+    const message = await prisma.message.findUnique({
+      where:{
+        id:messageId
+      }
+    });
+    if(!message){
+      return res.status(400).json({error:'no Message found'});
+    }
+    if(message.type==='file'){
+      await deleteObject(message.message);
+    }
 
-    const message = await prisma.message.delete({
+    await prisma.message.delete({
       where: {
         id: messageId,
       },
@@ -222,59 +294,4 @@ export async function downloadFileController(
   }
 }
 
-export async function uploadFileController(
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) {
-  try {
-    const file = req.file;
-    const senderId = req.query.senderId?.toString();
-    const receiverId = req.query.receiverId?.toString();
-    const type = req.file?.mimetype || "";
-    const filePath = req.file?.path || "";
 
-    if (!file) return res.status(400).json({ error: `File is required` });
-    if (!senderId)
-      return res.status(400).json({ error: `senderId is required` });
-    if (!receiverId)
-      return res.status(400).json({ error: `receiverId is required` });
-
-    const filename = file?.filename || "";
-    const key = `uploads/file-message/${Date.now()}-${filename}`;
-
-    const url = await putObject({
-      key,
-      contentType: type,
-    });
-
-    const response = await axios.put(url, file.buffer, {
-      headers: {
-        "Content-Type": type,
-      },
-    });
-    // console.log(response);
-
-    const imageUrl = await getObjectURL(key);
-
-    const message = await prisma.message.create({
-      data: {
-        senderUser: {
-          connect: { id: senderId },
-        },
-        receiverUser: {
-          connect: { id: receiverId },
-        },
-        message: imageUrl,
-        type: "file",
-      },
-    });
-
-    return res.status(201).json(imageUrl);
-  } catch (e) {
-    console.log(`Error in uploadFileController ${e}`);
-    return res
-      .status(500)
-      .json({ error: `Error in uploadFileController ${e}` });
-  }
-}
